feat(ui) Timestamp: Add UTC options parseDate and makeDate(Time)
Timestamp does not know anything about time zones, but when parsing
a Date object or creating one Timestamp has to know whether to use
local or utc time.

Before this, Timestamp read Date as local time but writes as UTC, so
makeDateTime(parseDate(x)) would create a Date with a different time.